Move prepConfig to root
- Since it's used by multiple commands
- Also clean up imports with goimports
